refactor: moved cards to components instead
